+++ src/lib.rs
@@ -38,3 +38,5 @@ pub mod consensus;
 
 /// tui Color assignments
 pub mod ui;
+
+pub mod runapp;